+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added new endpoints for creating papers and contributions.
   (Closes: [#279](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/issues/279),
   See: [!582](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/merge_requests/582))
+### Changed
+- Fetching several comparisons no longer returns previous version comparisons when no filter is set or filtering by visibility.
+  (See: [!630](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/merge_requests/630))
 ### Fixed
 - Lists that still have elements can now be deleted.
   (See: [!627](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/merge_requests/627))